多次评论时存在bug

+++ topic.php
@@ -62,19 +62,21 @@ else
 			
 				while($posts_row = mysqli_fetch_assoc($posts_result))
 				{
-					// echo "<style type='text/css'>a.delete-post{position:relative;bottom:-20px;right:-300px;}</style>";
+					echo "<style type='text/css'>a.delete-post{position:relative;bottom:-20px;right:-300px;}</style>";
 					echo '<tr class="topic-post"' .'id=' . $posts_row["post_id"] . '>' .
 							'<td class="user-post">' . $posts_row['user_name'] . '<br/>' . date('d-m-Y H:i', strtotime($posts_row['post_date'])) . '</td>
 							<td class="post-content">' . htmlentities(stripslashes($posts_row['post_content']));
-							if($_SESSION['user_id'] == $posts_row['user_id']){
+							echo "id是:" . $posts_row["post_id"] . "主题id是:" . $posts_row['post_topic'];
+							if (isset($_SESSION['user_id']) and ($_SESSION['user_id'] == $posts_row['user_id'])){
 								echo "<a class='delete-post' href='delete_post.php?post_id=" .$posts_row['post_id'].'&topic_id='.$posts_row['post_topic']. "'>删除</a>";
+								echo "<a class='update-post' href='update_post.php?post_id=" .$posts_row['post_id'].'&topic_id='.$posts_row['post_topic']. "'>修改</a>";
 							echo '</td>';
 					echo '</tr>';
 					}
 				}
 			}
 			
-			if(!$_SESSION['signed_in'])
+			if(!isset($_SESSION['signed_in']))
 			{
 				echo '<tr><td colspan=2>You must be <a href="signin.php">signed in</a> to reply. You can also <a href="signup.php">sign up</a> for an account.';
 			}
